added logic for undefined state docks
docks are now by default set to an invalid or undefined
state if not set otherwise. this means that if a any entity
tries to retrieve a value from a sink with get_value() it
will receive a NodeError if the sink is invalid. This way,
nodes that can't produce valid results if not all inputs are
connected or a specific set of settings is done in their config
ui can communicate this to nodes that depend on results from their
sources. Nodes that only generate values must be set valid by
hand. This can be accomplished via the new set_valid()-method